update packtypes.py; fixed bugs in socket; init c2pool/p2pool in message_version

+++ util/types.cpp
@@ -12,7 +12,7 @@ namespace c2pool::util::messages
         port = 0;
     }
 
-    address_type::address_type(int _services, std::string _address, int _port)
+    address_type::address_type(unsigned long long _services, std::string _address, int _port)
     {
         services = _services;
         address = _address;
